+++ spexxy/grid/__init__.py
@@ -4,3 +4,4 @@ from .values import ValuesGrid
 from .synspec import SynspecGrid
 from .ferre import FerreGrid
 from .fits import FitsGrid
+from .sub import SubGrid