+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v2.6
 * Added new FitsGrid, which stores all spectra in a single FITS file.
 * Added convert methods for new grid.
+* Added SubGrid and SubInterpolator, which both expose parts of a parent grid/interpolator.
 
 ### v2.5 (2020-07-14)
 * Split MainRoutine into MainRoutine and FilesRoutine, allowing for routines that don't run on files.